# parking/.junie/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/users/register:
    post:
      tags:
      - user-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
  /api/users/login:
    post:
      tags:
      - user-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/users/{id}:
    get:
      tags:
      - user-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
  /api/users/{id}/role:
    get:
      tags:
      - user-controller
      operationId: getUserRole
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                enum:
                - CITIZEN
                - ADMIN
...
  /api/reservations/users/{userId}:
    get:
      tags:
      - reservation-controller
      operationId: getReservationsForUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          maxLength: 2147483647
          minLength: 8
        numberplate:
          type: string
        role:
          type: string
          enum:
          - CITIZEN
          - ADMIN
      required:
      - email
      - numberplate
      - password
    User:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        email:
          type: string
        role:
          type: string
          enum:
          - CITIZEN
          - ADMIN
        numberplate:
          type: string
        createdAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
      - email
      - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    Reservation:
      type: object
      properties:
        reservationId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        spaceId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - PENDING
          - CONFIRMED
          - CANCELLED
          - COMPLETED
          - EXPIRED
        reservedFrom:
          type: string
          format: date-time
        reservedUntil:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...